+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # adrez-mapper Revisions
 
-202-03-17 (0.5.6)
+2020-03-23 (0.5.7)
+- campaign.action has now a standized lookup function
+
+
+2020-03-17 (0.5.6)
 - Campaign code / Code guid converted to typeGuid for definition
 
 2020-02-23 (0.5.5)